refactored parser to instead use magento API for addresses
