+++ web/app/connectors/_merlins-connector/checkout/commands.js
@@ -5,7 +5,7 @@
 import {makeJsonEncodedRequest, makeRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
 import {SubmissionError} from 'redux-form'
 import {createPropsSelector} from 'reselect-immutable-helpers'
-import {parseShippingInitialValues, parseLocations, parseShippingMethods, checkoutConfirmationParser, getNameValue} from './parsers'
+import {parseShippingInitialValues, parseLocations, parseShippingMethods, checkoutConfirmationParser} from './parsers'
 import {parseCartTotals} from '../cart/parser'
 import {parseCheckoutEntityID, extractMagentoShippingStepData} from '../../../utils/magento-utils'
 import {getCart} from '../cart/commands'
@@ -30,7 +30,10 @@ import * as shippingSelectors from '../../../store/checkout/shipping/selectors'
 import {getCartItemsFull} from 'progressive-web-sdk/dist/store/cart/selectors'
 import {getIsLoggedIn} from '../../../store/user/selectors'
 import {getShippingFormValues} from '../../../store/form/selectors'
-import {prepareEstimateAddress} from '../utils'
+import {
+    prepareEstimateAddress,
+    parseAddress
+} from '../utils'
 
 const INITIAL_SHIPPING_ADDRESS = {
     countryId: 'us',
@@ -69,36 +72,23 @@ export const fetchShippingMethodsEstimate = (inputAddress) => (dispatch, getStat
         })
 }
 
+export const fetchCustomerAddresses = () => {
+    const fetchURL = `/rest/default/V1/carts/mine`
+    return makeRequest(fetchURL, {method: 'GET'})
+        .then((response) => response.json())
+}
+
 export const fetchSavedShippingAddresses = (selectedSavedAddressId) => {
     return (dispatch) => {
-        const fetchURL = `/rest/default/V1/carts/mine`
-        return makeRequest(fetchURL, {method: 'GET'})
-            .then((response) => response.json())
+        fetchCustomerAddresses()
             .then(({customer}) => {
                 let defaultShippingId
                 const addresses = customer.addresses.map((address) => {
                     if (address.default_shipping) {
                         defaultShippingId = address.id
                     }
-                    const [addressLine1, addressLine2] = address.street
 
-                    // Not spreading `address` because it has key/values that
-                    // we want to rename and remove
-                    return {
-                        city: address.city,
-                        countryId: address.country_id,
-                        id: `${address.id}`,
-                        firstname: address.firstname,
-                        lastname: address.lastname,
-                        fullname: getNameValue(address.firstname, address.lastname),
-                        postcode: address.postcode,
-                        regionId: `${address.region.region_id}`,
-                        region: address.region.region,
-                        regionCode: address.region.region_code,
-                        addressLine1,
-                        addressLine2,
-                        telephone: address.telephone,
-                    }
+                    return parseAddress(address)
                 })
 
                 dispatch(setDefaultShippingAddressId(selectedSavedAddressId || defaultShippingId))
